# config/openapi/stackdome_api.yaml
openapi: 3.0.0
info:
  title: stackdome-api-server
  version: 0.0.1
  description: stackdome-api-server
paths:
...
  /api/v1/organizations/{org_id}/teams/{team_name}/stacks/{id}/releases:
    post:
      summary: Create a new release (deploy)
      operationId: createRelease
      security:
        - Bearer: []
      tags:
        - Releases
      parameters:
        - $ref: '#/components/parameters/org_id'
        - $ref: '#/components/parameters/team_name'
        - $ref: '#/components/parameters/id'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReleaseRequest'
      responses:
        '201':
          description: Release created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StackRelease'
    get:
      summary: List releases for a stack
      operationId: listReleases
      security:
        - Bearer: []
      tags:
        - Releases
      parameters:
        - $ref: '#/components/parameters/org_id'
        - $ref: '#/components/parameters/team_name'
        - $ref: '#/components/parameters/id'
        - in: query
          name: state
          schema:
            $ref: '#/components/schemas/StackReleaseState'
          description: Filter by release state
        - in: query
          name: page
          schema:
            type: integer
            default: 1
          description: Page number
        - in: query
          name: page_size
          schema:
            type: integer
            default: 20
          description: Number of items per page
      responses:
        '200':
          description: List of releases
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StackReleaseList'
  /api/v1/organizations/{org_id}/teams/{team_name}/stacks/{id}/releases/{release_id}:
    get:
      summary: Get a release by ID
      operationId: getRelease
      security:
        - Bearer: []
      tags:
        - Releases
      parameters:
        - $ref: '#/components/parameters/org_id'
        - $ref: '#/components/parameters/team_name'
        - $ref: '#/components/parameters/id'
        - name: release_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Release details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StackReleaseDetail'
  /api/v1/organizations/{org_id}/teams/{team_name}/stacks/{id}/releases/{release_id}/cancel:
    post:
      summary: Cancel a pending or rendering release
      operationId: cancelRelease
      security:
        - Bearer: []
      tags:
        - Releases
      parameters:
        - $ref: '#/components/parameters/org_id'
        - $ref: '#/components/parameters/team_name'
        - $ref: '#/components/parameters/id'
        - name: release_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Release cancelled

  # Team-scoped secret routes
...
components:
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    id:
      name: id
      in: path
      description: The id of record
      required: true
      schema:
        type: string
    org_id:
      name: org_id
      in: path
      required: true
      schema:
        type: string
      description: The ID of the organization
...
    team_name:
      name: team_name
      in: path
      required: true
      schema:
        type: string
      description: The name of the team
...
  schemas:
...
    Volume:
      type: object
      additionalProperties: false
      required:
        - name
        - spec
      properties:
        id:
          type: string
          readOnly: true
        team_id:
          type: string
          readOnly: true
        name:
          type: string
        labels:
          type: array
          items:
            $ref: '#/components/schemas/Label'
        annotations:
          type: array
          items:
            $ref: '#/components/schemas/Annotation'
        spec:
          $ref: '#/components/schemas/VolumeSpec'
        status:
          $ref: '#/components/schemas/VolumeStatus'
          readOnly: true
...
    VolumeSpec:
          type: object
          required:
            - size
            - needs_sync_before_use
            - access_mode
          additionalProperties: false
          properties:
            size:
              type: string
            storage_class:
              type: string
            needs_sync_before_use:
              type: boolean
            access_mode:
              $ref: '#/components/schemas/VolumeAccessMode'
            source:
              $ref: '#/components/schemas/VolumeSource'
    VolumeAccessMode:
      type: string
      enum:
        - ReadWriteOnce
        - ReadWriteMany
        - ReadOnlyMany
    VolumeSource:
      type: object
      required:
        - source_type
      additionalProperties: false
      properties:
        git_repo_source:
          $ref: '#/components/schemas/GitRepoSource'
        source_type:
          $ref: '#/components/schemas/VolumeSourceTypes'
        remote_source:
          $ref: '#/components/schemas/RemoteSource'
        build_source:
          type: array
          items:
            $ref: '#/components/schemas/BuildArtifact'
    VolumeSourceTypes:
      type: string
      enum:
      - RemoteDir
      - BuildArtifact
      - GitRepo
    GitRepoSource:
      type: object
      required:
        - repo_url
        - revision
      properties:
        repo_url:
          type: string
        revision:
          $ref: '#/components/schemas/GitRepoRevision'
    GitRepoRevision:
      type: object
      properties:
        branch:
          type: string
        tag:
          type: string
        commit:
          type: string
    RemoteSource:
      type: object
      required:
        - path
        - current_directory_hash
      properties:
        path:
          type: string
        current_directory_hash:
          type: string
    BuildArtifact:
      type: object
      required:
        - resource_ref
        - source_path
        - destination_path
      properties:
        resource_ref:
          type: string
        source_path:
          type: string
        destination_path:
          type: string
    VolumeStatus:
      type: object
      properties:
        conditions:
          type: array
          items:
            $ref: '#/components/schemas/Condition'
        phase:
          type: string
        build_artifact_syncs:
          type: array
          items:
            $ref: '#/components/schemas/BuildArtifactSyncInfo'
        last_synced_git_revision:
          type: string
        last_remote_sync_hash:
          type: string
    BuildArtifactSyncInfo:
      type: object
      properties:
        resource_name:
          type: string
        build_id:
          type: string
        status:
          type: string
...
    StackResource:
      type: object
      properties:
        id:
          type: string
          readOnly: true
        stack_id:
          type: string
          readOnly: true
        name:
          type: string
        labels:
          type: array
          items:
            $ref: '#/components/schemas/Label'
        annotations:
          type: array
          items:
            $ref: '#/components/schemas/Annotation'
        revision:
          type: string
          readOnly: true
        source:
          $ref: '#/components/schemas/SourceSpec'
        init_spec:
          $ref: '#/components/schemas/InitSpec'
        execution_config:
          $ref: '#/components/schemas/ExecutionConfig'
        volume_mounts:
          type: array
          items:
            $ref: '#/components/schemas/VolumeMount'
        depends_on:
          type: array
          items:
            type: string
        lifecycle_config:
          $ref: '#/components/schemas/LifecycleConfig'
        ports:
          type: array
          items:
            $ref: '#/components/schemas/Port'
        outputs:
          type: array
          readOnly: true
          items:
            $ref: '#/components/schemas/OutputDescriptor'
        workload_type:
          type: string
          enum:
            - Service
            - StatefulService
            - Worker
            - Job
            - CronJob
          default: Service
        schedule:
          type: string
        replicas:
          type: integer
          format: int32
          minimum: 0
        status:
          $ref: '#/components/schemas/StackResourceStatus'
          readOnly: true
      required:
        - name
...
    OutputDescriptor:
      type: object
      description: Declared output metadata for a topology node. Values are not returned here.
      required:
        - name
        - type
        - sensitive
      properties:
        name:
          type: string
          description: Stable output accessor name, for example `host` or `public.http.url`.
        type:
          type: string
          description: Scalar value type exposed by this output.
          enum:
            - string
            - integer
            - boolean
        sensitive:
          type: boolean
          description: True when the output value is sensitive and should never be returned in normal metadata APIs.
...
    StackConnection:
      type: object
      description: A user-authored topology edge between two nodes in a stack.
      required:
        - kind
        - from
        - to
      properties:
        id:
          type: string
          description: Stable connection identifier. Generated when omitted.
        kind:
          type: string
          description: >
            The relationship type. `env` injects values into environment variables,
            `volume_mount` mounts a volume into a resource, and
            `build_artifact_source` seeds a volume from build output.
          enum:
            - env
            - volume_mount
            - build_artifact_source
        from:
          $ref: '#/components/schemas/TopologyNodeRef'
        to:
          $ref: '#/components/schemas/TopologyNodeRef'
        mappings:
          type: array
          description: >
            Target/value mappings for kinds that move values, such as `env`.
          items:
            $ref: '#/components/schemas/ConnectionMapping'
        config:
          $ref: '#/components/schemas/StackConnectionConfig'
    TopologyNodeRef:
      type: object
      description: Identifies a topology node within the stack graph.
      required:
        - type
      properties:
        type:
          type: string
          description: The node category.
          enum:
            - stack_resource
            - addon/postgres
            - secret
            - volume
            - object_store
        id:
          type: string
          description: Stable ID for persisted resources such as addons or secrets.
        name:
          type: string
          description: Name-scoped reference for stack-local resources such as StackResources or Volumes.
    ConnectionMapping:
      type: object
      description: Maps one produced value into one target location on the consumer.
      required:
        - target
        - value
      properties:
        target:
          $ref: '#/components/schemas/ConnectionTarget'
        value:
          $ref: '#/components/schemas/ValueRef'
    ConnectionTarget:
      type: object
      description: The destination field that receives a mapped value.
      required:
        - type
      properties:
        type:
          type: string
          description: env writes an environment variable and file writes a mounted file path.
          enum:
            - env
            - file
        name:
          type: string
          description: Environment variable name when type is env.
        path:
          type: string
          description: Absolute file path when type is file.
    StackConnectionConfig:
      description: >
        Kind-specific connection configuration. The shape depends on the
        connection kind and source type: use PostgresEnvConfig when kind is
        env and from.type is addon/postgres, VolumeMountConfig when kind is
        volume_mount, and BuildArtifactSourceConfig when kind is
        build_artifact_source. Omit config entirely for env connections from
        stack_resource or secret sources.
      oneOf:
        - $ref: '#/components/schemas/PostgresEnvConfig'
        - $ref: '#/components/schemas/VolumeMountConfig'
        - $ref: '#/components/schemas/BuildArtifactSourceConfig'
    PostgresEnvConfig:
      type: object
      description: >
        Config for env connections from a PostgreSQL addon
        (kind=env, from.type=addon/postgres). credential_scope and superuser
        are mutually exclusive.
      properties:
        database:
          type: string
          description: Target database name within the addon.
        credential_scope:
          type: string
          enum:
            - owner
            - superuser
          description: Which credential set to inject. Mutually exclusive with superuser.
        superuser:
          type: boolean
          description: Use superuser credentials. Mutually exclusive with credential_scope.
      additionalProperties: false
    VolumeMountConfig:
      type: object
      description: Config for volume mount connections (kind=volume_mount).
      required:
        - mount_path
      properties:
        mount_path:
          type: string
          description: Absolute path where the volume is mounted in the container.
        sub_path:
          type: string
          description: Sub-path within the volume to mount.
        read_only:
          type: boolean
          description: Mount the volume read-only.
      additionalProperties: false
    BuildArtifactSourceConfig:
      type: object
      description: Config for build artifact source connections (kind=build_artifact_source).
      required:
        - source_path
      properties:
        source_path:
          type: string
          description: Path within the build output to copy from.
        destination_path:
          type: string
          description: Path within the volume to copy to.
      additionalProperties: false
    ValueRef:
      type: object
      description: >
        Describes how to read a value from the connection's `from` node. This is
        only used inside `StackConnection.mappings[]`.
      properties:
        output:
          type: string
          description: Output accessor on the connection's `from` node, such as `url` or `public.http.url`.
        template:
          type: string
          description: Template used when one target value must be composed from multiple outputs.
        values:
          type: object
          description: Named template inputs, each resolving one output from the connection's `from` node.
          additionalProperties:
            $ref: '#/components/schemas/OutputValueRef'
    OutputValueRef:
      type: object
      description: References one output on the connection's `from` node.
      required:
        - output
      properties:
        output:
          type: string
          description: Output accessor on the connection's `from` node.
...
    StackResourceStatus:
      type: object
      properties:
        public_ingress:
          type: array
          items:
            $ref: '#/components/schemas/Ingress'
        internal_service_name:
          type: string
        last_restart_request_processed_at:
          type: string
          format: date-time
        state:
          type: string
        observed_revision:
          type: string
        conditions:
          type: array
          items:
            $ref: '#/components/schemas/Condition'
        last_failure:
          $ref: '#/components/schemas/StackResourceFailure'
        replicas:
          type: integer
          format: int32
          readOnly: true
        available_replicas:
          type: integer
          format: int32
          readOnly: true
        updated_replicas:
          type: integer
          format: int32
          readOnly: true
        last_run_time:
          type: string
          format: date-time
          readOnly: true
        last_run_succeeded:
          type: boolean
          readOnly: true
    ContainerFailureDetail:
      type: object
      properties:
        failure_type:
          type: string
          enum:
            - crash_loop
            - out_of_memory
            - image_pull_failed
            - create_container_error
            - exit_error
        reason:
          type: string
        message:
          type: string
        restart_count:
          type: integer
          format: int32
        exit_code:
          type: integer
          format: int32
    BuildFailureDetail:
      type: object
      properties:
        failure_type:
          type: string
          enum:
            - crash_loop
            - out_of_memory
            - image_pull_failed
            - create_container_error
            - exit_error
        reason:
          type: string
        message:
          type: string
        restart_count:
          type: integer
          format: int32
        exit_code:
          type: integer
          format: int32
    StackResourceFailure:
      type: object
      properties:
        type:
          type: string
          enum:
            - runtime_crash
            - build_failure
        container:
          $ref: '#/components/schemas/ContainerFailureDetail'
        init_container:
          $ref: '#/components/schemas/ContainerFailureDetail'
        build:
          $ref: '#/components/schemas/BuildFailureDetail'
    Ingress:
      type: object
      properties:
        url:
          type: string
        target_port:
          type: integer
    VolumeMount:
      type: object
      required:
        - source_volume_name
        - target_path
      properties:
        stack_resource_id:
          type: string
          readOnly: true
        source_volume_type:
          $ref: '#/components/schemas/VolumeMountSourceType'
          readOnly: true
        source_volume_name:
          type: string
        source_sub_path:
          type: string
        target_path:
          type: string
    VolumeMountSourceType:
      type: string
      enum:
      - EmptyVolume
      - RemoteDirSyncedVolume
      - BuildArtifactSyncedVolume
      - GitRepoSyncedVolume
    Port:
      type: object
      required:
        - name
        - number
        - exposed_to_public
      properties:
        name:
          type: string
        number:
          type: integer
        protocol:
          type: string
        exposed_to_public:
          type: boolean
        subdomain_prefix:
          type: string
    LifecycleConfig:
      type: object
      properties:
        restart_request_time:
          type: string
          format: date-time
    SourceSpec:
      type: object
      additionalProperties: false
      description: Where a stack resource's runtime image comes from. Exactly one of git, image, or volume must be set (enforced server-side).
      properties:
        git:
          $ref: '#/components/schemas/GitSource'
        image:
          $ref: '#/components/schemas/ImageSource'
        volume:
          $ref: '#/components/schemas/VolumeBuildSource'
    GitSource:
      type: object
      additionalProperties: false
      required:
        - repo_url
      properties:
        repo_url:
          type: string
        branch:
          type: string
          description: Defaults to the repository's default branch, resolved and stored at create time
        tag:
          type: string
          description: Mutually exclusive with branch
        commit:
          type: string
          description: Commit SHA pin; requires branch or tag
        dockerfile_path:
          type: string
          default: Dockerfile
        build_context:
          type: string
          default: "."
        integration_id:
          type: string
          description: Org-level git integration override for clone auth
        push:
          $ref: '#/components/schemas/PushTarget'
          description: Omit to push to the internal cluster registry
    PushTarget:
      type: object
      additionalProperties: false
      required:
        - repository
      properties:
        repository:
          type: string
          description: Push repository as host/path, without a tag
        registry_credentials_id:
          type: string
          description: Org-level registry credential override for push auth
    ImageSource:
      type: object
      additionalProperties: false
      required:
        - ref
      properties:
        ref:
          type: string
        registry_credentials_id:
          type: string
          description: Org-level registry credential override for pull auth
    VolumeBuildSource:
      type: object
      additionalProperties: false
      properties:
        volume_id:
          type: string
        volume_name:
          type: string
          description: Name of a volume defined on the stack; either volume_id or volume_name is required
        current_volume_hash:
          type: string
          description: Content hash of the volume used as the build source revision
        dockerfile_path:
          type: string
          default: Dockerfile
        build_context:
          type: string
          default: "."
...
    InitSpec:
      type: object
      additionalProperties: false
      properties:
        command:
          type: array
          items:
            type: string
        args:
          type: array
          items:
            type: string
    ExecutionConfig:
      type: object
      properties:
        command:
          type: array
          items:
            type: string
        args:
          type: array
          items:
            type: string
        environment_variables:
          type: array
          items:
            $ref: '#/components/schemas/EnvVar'
    EnvVar:
      type: object
      additionalProperties: false
      required:
        - name
      properties:
        name:
          type: string
        value:
          type: string
          description: Literal environment variable value.
        self_output:
          type: string
          description: Read this environment variable from one of the resource's own declared outputs, for example public.http.url.
    Condition:
      type: object
      additionalProperties: false
      properties:
        type:
          type: string
        status:
          type: string
        observed_generation:
          type: integer
        last_transition_time:
          type: string
          format: date-time
        reason:
          type: string
        message:
          type: string
    Label:
      type: object
      additionalProperties: false
      required:
        - key
        - value
      properties:
        key:
          type: string
        value:
          type: string
    Annotation:
      type: object
      required:
        - key
        - value
      properties:
        key:
          type: string
        value:
          type: string
...
    StackReleaseState:
      type: string
      enum: [Pending, InProgress, Released, Failed, Superseded, Cancelled]
      x-enum-varnames:
        - RELEASE_STATE_PENDING
        - RELEASE_STATE_IN_PROGRESS
        - RELEASE_STATE_RELEASED
        - RELEASE_STATE_FAILED
        - RELEASE_STATE_SUPERSEDED
        - RELEASE_STATE_CANCELLED
    ReleaseCauseKind:
      type: string
      enum: [manual, rollback, webhook_push]
      x-enum-varnames:
        - RELEASE_CAUSE_MANUAL
        - RELEASE_CAUSE_ROLLBACK
        - RELEASE_CAUSE_WEBHOOK_PUSH
    CreateReleaseRequest:
      type: object
      properties:
        from_release_id:
          type: string
          description: If set, creates a rollback release copying this release's manifest
    StackRelease:
      type: object
      properties:
        id:
          type: string
        stack_id:
          type: string
        sequence:
          type: integer
        state:
          $ref: '#/components/schemas/StackReleaseState'
        message:
          type: string
        cause:
          $ref: '#/components/schemas/ReleaseCause'
        snapshot_revision:
          type: string
        manifest_revision:
          type: string
        renderer_version:
          type: string
        pins:
          $ref: '#/components/schemas/ReleasePins'
        outcome:
          $ref: '#/components/schemas/ReleaseOutcome'
        created_by:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        rendered_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
    StackReleaseDetail:
      allOf:
        - $ref: '#/components/schemas/StackRelease'
        - type: object
          properties:
            snapshot:
              $ref: '#/components/schemas/StackReleaseSnapshot'
    StackReleaseSnapshot:
      type: object
      properties:
        stack:
          type: object
          properties:
            id:
              type: string
            organisation_id:
              type: string
            team_id:
              type: string
            cluster_id:
              type: string
            user_id:
              type: string
            name:
              type: string
            namespace_id:
              type: string
            namespace:
              type: string
            labels:
              type: object
              additionalProperties:
                type: string
            annotations:
              type: object
              additionalProperties:
                type: string
        resources:
          type: array
          items:
            $ref: '#/components/schemas/StackResource'
        volumes:
          type: array
          items:
            $ref: '#/components/schemas/Volume'
        connections:
          type: array
          items:
            $ref: '#/components/schemas/StackConnection'
        captured_at:
          type: string
          format: date-time
    StackReleaseList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/StackRelease'
        total:
          type: integer
          description: Total number of records
        page:
          type: integer
          description: Current page number
        page_size:
          type: integer
          description: Number of items per page
        total_pages:
          type: integer
          description: Total number of pages
    ReleaseCause:
      type: object
      properties:
        kind:
          $ref: '#/components/schemas/ReleaseCauseKind'
        detail:
          type: string
    ReleasePins:
      type: object
      properties:
        resources:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/ResourcePins'
    ResourcePins:
      type: object
      properties:
        git_sha:
          type: string
        volume_hash:
          type: string
        image_digest:
          type: string
    ReleaseOutcome:
      type: object
      properties:
        resources:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/ResourceOutcome'
        duration:
          type: string
    ResourceOutcome:
      type: object
      properties:
        phase:
          type: string
        ready_replicas:
          type: integer
        replicas:
          type: integer
        message:
          type: string
...